+++ xugrid/ugrid/ugridbase.py
@@ -129,7 +129,7 @@ class AbstractUgrid(abc.ABC):
         pass
 
     @abc.abstractmethod
-    def sel_points(self):
+    def sel_points(self, obj, x, y, out_bounds, fill_value):
         pass
 
     @abc.abstractmethod